// src/app/core/enums/resume-replacement.enum.ts
export enum ResumeReplacementErrorCode {
  UPGRADE_REQUIRED = 'UPGRADE_REQUIRED',
  REPLACEMENT_QUOTA_EXCEEDED = 'REPLACEMENT_QUOTA_EXCEEDED',
  NO_ACTIVE_RESUME = 'NO_ACTIVE_RESUME',
  SAME_FILE_AS_ACTIVE = 'SAME_FILE_AS_ACTIVE',
  INVALID_FILE = 'INVALID_FILE',
  STORAGE_UPLOAD_FAILED = 'STORAGE_UPLOAD_FAILED',
  RESUME_PROFILE_NOT_READY = 'RESUME_PROFILE_NOT_READY',
  PROFILE_ENRICHMENT_IN_PROGRESS = 'PROFILE_ENRICHMENT_IN_PROGRESS',
  CANNOT_DELETE_ACTIVE_RESUME = 'CANNOT_DELETE_ACTIVE_RESUME',
  RESTORE_OUT_OF_WINDOW = 'RESTORE_OUT_OF_WINDOW',
  RESTORE_TARGET_NOT_FOUND = 'RESTORE_TARGET_NOT_FOUND',
  REPLACEMENT_TX_FAILED = 'REPLACEMENT_TX_FAILED',
  IDEMPOTENT_REPLAY = 'IDEMPOTENT_REPLAY',
}